+++ relecov_dashboard/views.py
@@ -48,6 +48,7 @@ def variants_lineage_variation_over_time(request):
 
 def variants_mutations_in_lineages_heatmap(request):
     create_hot_map(2018185)
+    # create_hot_map()
     return render(request, "relecov_dashboard/variantsMutationsInLineagesHeatmap.html")
